[frontend] Move TextField component to TypeScript
